Adding new booster pngs
Created using an online editor tool called Pixlr.com.
Paste the image as layer 1 and use their Remove Background AI tool. If it removes the inner background of the booster as well, add a second layer and the same unaltered image and use the cutout feature. manually cutout all but the middle portion of the booster on this layer. After this, when you combine both layers, you get images perfect.